requirements.txt
joblib
pydub==0.23.1
numpy==1.15.1
baidu-aip==2.2.13.0
pyltp==0.2.1
py2neo==4.2.0
flask==1.0.2
requests==2.20.1
tensorflow==1.13.1
itchat==1.3.10
